web service module integrated. It works !!

+++ src/interfaces/NetworkServiceInterface.java
@@ -438,7 +438,7 @@ public interface NetworkServiceInterface
      * @return the last update time about account personnal information .
      * @throws NetworkServiceException
      */
-    public int getLastPersonnalInformationUpdateTime() throws NetworkServiceException, NotAuthenticatedException;
+    public long getLastPersonnalInformationUpdateTime() throws NetworkServiceException, NotAuthenticatedException;
 
     /**
      *
